Adding color schemes to emacs
Prior to this commit, I was noticing odd behavior in the M-x selected
text. The contrast was off; The text was bold red and the background
box was green. This made things hard to read.

I tinkered around with settings of iTerm, zsh, and Emacs. This, along
with disabling the UI.zsh line seems to have stabilized behavior.

However, only time will tell.