Adding `e' alias for editor and tweaking VISUAL
I spent a bit poking around in Spacemacs, and I certainly see the
appeal. However it feels very heavy compared to emacs-plus. As I poke
around with both options, I need a quick switch between the two modes.

This is not that switch but captures some of what must change.