Removing mode-line moody + intense-hl modus config
I'm not using https://github.com/tarsius/moody, so it's appears to be
fiddling with Spaceline.  I tried switching to moody but it did not
include a Flycheck error section.  I really like having the Flycheck
error count in the modeline.

In the case of theme-intense-hl, the default appears to be nil, so no
sense declaring it here.