Refactoring gorg macro
Prior to this commit, I invoked the macro twice per element; Once for
the method definition and once for the keyboard binding.

With this change, I consolidated into one macro invocation per element.

Key is using the `progn` to indicate a list of function calls to
evaluate.

This is clearly a vanity affair.